packages/api/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stratmaster-api"
version = "0.1.0"
description = "StratMaster FastAPI service"
requires-python = ">=3.11"
authors = [{ name = "StratMaster" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "httpx>=0.24",
  "PyYAML>=6.0",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
